Fix a typo.
Do not write the same value to uartlcr_h and uartdrvout. A 0 should by written
to uarddrvout instead. This is not a fatal error as the value written to
uartdrvout is MSK_uartlcr_h_FEN | MSK_uartlcr_h_WLEN = 0x70. All these bits
are reserved in the uartdrvout register, so writing 0x70 there has the same
effect as writing a 0.